Extract Solaris CentralProcessor common base
Move the command-based logic into an abstract SolarisCentralProcessor base
in oshi-common: NUMA mapping, processor-ID feature flags, the dmesg
physical-processor parsing (via an abstract queryLogicalProcessors hook),
the averaged tick rollup, and the kstat -p context-switch/interrupt
fallbacks. The kstat-chain queries stay in the subclasses;
SolarisCentralProcessorJNA overrides queryContextSwitches/queryInterrupts to
add the Kstat2 path and delegates to super for the command fallback.
